댓글 수정 카카오로그인 수정, 게시글 글자수수정

+++ front/src/pages/board/BoardDetail.js
@@ -216,6 +216,10 @@ const BoardDetail = () => {
                         <div style={{ flex: 1 }}>
                           <CustomBiTrash
                             onClick={() => {
+                              if (user?.id !== item.UserId) {
+                                alert('자신이 쓴 댓글만 가능합니다.');
+                                return;
+                              }
                               deleteComment(item?.id, item);
                             }}
                           />
